fix: Update TooltipPop
